Create click handlers for new operators
The newly created operators were not being represented in the
JS to be handled when they are clicked, so they were added to
the existing #operators a click handler. However, since decimal
is not an actual operator and was added to the .not exclusion in
the event handler. An if/else statement had to be added to the
handler for the square root operator since it does not require a
second number to be inputted. Instead in this case the operator
was set to sqrt and the #equals click handler was triggered.