Click handler for #decimal
The #decimal div was excluded from the #operators div click handler
so it needed its own handler for its specific case. The handler was
added and first runs a for loop for all of the characters in
number, when a decimal is found the var numOfDecs is increased by
one. After the for loop an if statement is checked for numOfDecs
equalling 0, if so then the decimal is added to the end of number.
At the end of the click handler the text of totaldiv is changed to
the number and testNumLength is checked for number.