A document.ready to load jQuery/JS
The JS file did not begin with a document.ready function which meant
that the JS was not being loaded/activated properly in the browser.
Therefore a document.ready function statement was added at the top
of the file and included all of the defined functions that were
already within the file. All of the event handlers now operate
properly.