Update colors for calculator
The original calculator was very bland and not my taste, so I
decided to add a little color to it to make it more exciting. The
background color for the whole calculator had already been changed
to a dark turquoise, but they whole color pallette didn't work out.
Therefore I switched it out for a bright shade of pink. The
background colors of the numbers were changed to a light pink with
dark blue font. The operators were changed to a dark purple with
white font and the equals button was changed to navy instead of
purple.